move and rename files, fix comments
